I added a rule to highlight URLs because someone asked about it on the ML, and it turned out to actually be quite nice to have them underlined, so now I'm committing it.
The pattern used is a huge simplification, but all patterns I found which allow only valid URLs (to some degree) doesn't handle when the URL is e.g. in (parenthesis), "quoted", followed by a comma, or similar. But anyone wanting to make the pattern more complex is of course welcome, just remember to handle these cases.


git-svn-id: http://svn.textmate.org/trunk/Bundles/Text Utilities.tmbundle@1071 dfb7d73b-c2ec-0310-8fea-fb051d288c6d